pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moss-hermes"
version = "0.1.0"
description = "MOSS integration for Nous Research Hermes agents - cryptographic signing for AI actions"
readme = "README.md"
license = "MIT"
authors = [
    { name = "MOSS Computing", email = "support@mosscomputing.com" }
]
keywords = [
    "moss",
    "hermes",
    "nous-research",
    "ai-governance",
    "cryptographic-signing",
    "ml-dsa-44",
    "post-quantum",
    "ai-agents",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "moss-sdk>=0.2.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://mosscomputing.com"
Documentation = "https://docs.mosscomputing.com/hermes"
Repository = "https://github.com/mosscomputing/moss-hermes"
Changelog = "https://github.com/mosscomputing/moss-hermes/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["moss_hermes"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
